Lesson 2 - Level 3 - Task 11

+++ src/main/java/students/vladimir_aboltins/lesson_2/level_3/Task_11/MeanValueOfThree.java
@@ -12,11 +12,11 @@ public class MeanValueOfThree {
         greatings();
 
         printAskFirstNumber();
-        int firstNumber = getNumFromUser();// Ввод первого числа
+        float firstNumber = getNumFromUser();// Ввод первого числа
         printAskSecondNumber();
-        int secondNumber = getNumFromUser();// Ввод второго числа
+        float secondNumber = getNumFromUser();// Ввод второго числа
         printAskThirdNumber();
-        int thirdNumber = getNumFromUser();// Ввод третего числа
+        float thirdNumber = getNumFromUser();// Ввод третего числа
         float valueNumber=(firstNumber+secondNumber+thirdNumber)/3;
         System.out.println("Mean value of three is = " + valueNumber);
